[WIP] Properly support multi-world difficulties.
This adds a "difficulty" field to the World section of the Sponge world
config files. If this field is not present, empty, or null then we use
the World difficulty. If the multi-world command mixin is enabled then
calling difficulty will set it for the world you are in and write that
value to the config. Only when a mod calls setDifficulty of a WorldInfo,
a plugin via WorldProperties, or via the command will update the config.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>